// src/app/model/pull-request.ts
export interface IPullRequestRoot {
    success: boolean
    data: IPullRequest[]
  }
  
  export interface IPullRequest {
    url: string
    id: number
    node_id: string
    html_url: string
    diff_url: string
    patch_url: string
    issue_url: string
    number: number
    state: string
    locked: boolean
    title: string
    user: User
    body: any
    created_at: string
    updated_at: string
    closed_at: any
    merged_at: any
    merge_commit_sha: string
    assignee: any
    assignees: any[]
    requested_reviewers: any[]
    requested_teams: any[]
    labels: any[]
    milestone: any
    draft: boolean
    commits_url: string
    review_comments_url: string
    review_comment_url: string
    comments_url: string
    statuses_url: string
    head: Head
    base: Base
    _links: Links
    author_association: string
    auto_merge: any
    active_lock_reason: any
  }
  
  export interface User {
    login: string
    id: number
    node_id: string
    avatar_url: string
    gravatar_id: string
    url: string
    html_url: string
    followers_url: string
    following_url: string
    gists_url: string
    starred_url: string
    subscriptions_url: string
    organizations_url: string
    repos_url: string
    events_url: string
    received_events_url: string
    type: string
    user_view_type: string
    site_admin: boolean
  }
  
  export interface Head {
    label: string
    ref: string
    sha: string
    user: User2
    repo: Repo
  }
// ...
  export interface Base {
    label: string
    ref: string
    sha: string
    user: User3
    repo: Repo2
  }
// ...
  export interface Links {
    self: Self
    html: Html
    issue: Issue
    comments: Comments
    review_comments: ReviewComments
    review_comment: ReviewComment
    commits: Commits
    statuses: Statuses
  }